// types/sidebar-items.ts
export enum SidebarItems {
    PRODUCTS = "Products",
    ORDERS = "Orders",
    SALES = "Sales",
    INVOICES = "Invoices",
    TRANSACTIONS = "Transactions",
    INVENTORY = "Inventory",
    FLASH_SALES = "Flash Sales",
    COUPONS = "Coupons",
    BLOGS = "Blogs",
    CONTACTS = "Contacts",
    PARTNER_COMPANY = "Partner Company",
    FAQS = "FAQ's",
    NEWSLETTERS = "Newsletters",
    SMS = "SMS",
    BANNERS = "Banners",
    NAVIGATION_INFO = "Navigation Info",
    NOTIFICATION = "Notification",
    USER_MANAGEMENT = "User Management",
    TESTIMONIALS = "Testimonials",
    EXPERT_RECOMMENDATIONS = "Expert Recommendations",
    CAREERS = "Careers",
    SETTINGS = "Settings",
    REPORTS = "Reports",
    DASHBOARD = "Dashboard",
    CANCEL_ITEMS = "Cancel Items",
    RETURN_ITEMS = "Return Items",
    NEWSLETTERS_CLIENTS = "Newsletter Client",
    TEAM = "Team",
    ADVERTISMENT = "Advertisement Banner",
    // COMPARE = "Comparison"
}